refactor: Change the main layout
Add buttons "sign.in" and "sign up" instead of a choice box in the MainMenuButtons.fxml.
Add their keys to the bundle "messages". Change buttons' location by adding 2 VBox and 2 HBox.
Move the imported MainMenuButtons.fxml in the MainLayout.fxml from the center to the top. Delete the button exit.