refactor: Change the project layout
Change the project buttons' orientation from horizontal to vertical (VBox instead of HBox)
in the ProjectButtons.fxml. Change the location of the ProjectButtons.fxml in the MainProject.fxml from
the bottom to the left. Change the location of the choice box from the MainProject.fxml to the ProjectMenu.fxml.
Resize contents of ProjectMenu.fxml. Delete the MenuBar in the MainProject.fxml.