feat: Add the choose.button's action - upload a user photo
Change project buttons' location from the projectButtons.fxml to the ImageView.fxml.
Delete the projectButtons.fxml, the userChoice.fxml and their controllers. Add the class
ImageFX with the SimpleStringProperty (photo name) and the ObjectProperty containing a photo.
Add the class ImageFxModel with the method loadImage which returns the image uploaded
from computer files. Create the choose.button's action method "choosePhoto" in the
ImageViewController which set the image on the left ImageView.