refactor: Change the way of loading fxml files
Create a class "App" which contains a reference to the parent BorderPane
and a static method setCenterLayout setting fxml files to the center part of BorderPane.
Make the primaryStage and the mainPane static in the class Main. Refactor the method
openProject in the MainMenuButtonsController so as to use the setCenterLayout method.