feat: Create disableProperty bindings
Create disableProperty bindings in the ImageViewController to make the
createProjectButton and the deleteButton disabled when the imageBefore is not set
and the zoomButton and the saveAsButton disabled when the imageAfter is not created.
Create a binding between the saveNameButton and the projectNameField in the
ProjectMenuController so as to make the button disabled when the textField is empty.