feat: Add updating readyImage with parameter set by slider
Add a slider that sets parameter used in the dilation algorithm.
Initialize the slider in the initialize method in the class
MainProjectController. Add dilationValue in the class ImageFxProject.
Add a method dilate in the class ImageFxProjectModel that create and
set image updated with parameter. Add dilation in the method detectEdges.