feat: Add making black and white photo with a parameter set by slider
Add the slider that sets the parameter used in determining contrast in a photo.
Initialize the slider in the initialize method in the MainProjectController.
Add contrastValue (double) in the ImageFxProject. Add a method makeContrast
in the ImageFxProjectModel that updates readyImage using a new value of the parameter.
Add a method contrast in the class EdgeDetection which makes pixels black (color above the parameter)
or white (color below the parameter) to make black and white photo instead of a gray one.