feat: Add opening zoomWindow on a button click
Create a fxml file ZoomWindow with an ImageView that shows the readyImage
and a ZoomController with a method that updates photo in the ImageView.
Add a method openZoom in the MainProjectController. Add a method setZoom that
is called in the ProjectModel initialization method and creates stage using FXMLLoader.
Add a method showZoom that sets the readyImage and presents the window created in the
setZoom method. Add updating an image presented in the zoomWindow in methods
dilate and makeContrast.