refactor: Modify working with a list of projects
Create a class ProjectList that contains ObservableList with projects
and methods which works on that list. Create a class ProjectListModel
that contains an instance of ProjectList class and static methods
that let work on the list in the MainProjectController without creating
a class instance in the controller. Add calling a method addProject in the
method saveProject in the MainProjectController.