feat: Add opening a project selected in the ProjectChoiceComboBox
Add a listener to the ProjectChoicecomboBox in the initialize method of
MainProjectController that calls a method loadSelectedProject after
noticing a change. Add the method loadSelectedProject in the MainProjectController
that sets elements of the selected project. Add a method updateProject in the
ProjectModel that replace an activeProjext with the selected one.