feat: Add deleting project on a right mouse click
Add a method clickProjectChoiceComboBox that calls a static method
deleteProjectOnRightClick in the class ProjectListModel if the
ComboBox has been clicked with the right mouse button. Add a method
deleteProjectOnRightClick in the class ProjectListModel. Add a method
deleteProjectOnMouseEvent in the class ProjectList. Add a method
showDeleteConfirmationDialog that shows a confirmation dialog and calls
the method deleteProject if the answer is "ok". Add a method
showDeleteConfirmationDialog in the class DialogUtils. Update the
ProjectChoiceComboBox listener. Add a method loadBlankProject.