feat: Add removing a project from the database
Add a method deleteProject in the class ProjectDao that call a static
method executeDeletion with a proper SQL statement. Add methods
deleteProject and createDeleteStatement in the class DbManager. Add
the method exectuteDelation that use methods deleteProject and
createDeleteStatement to make database connection and delete the given
project. Add deleting a project in the class ProjectList.