feat: Add updating projects in the database
Add a method updateProject in the class ProjectDao that call a static
method executeProjectUpdate with a proper SQL statement. Add methods
updateProject and createUpdateStatement in the class DbManager. Add
the method executeProjectUpdate that use methods updateProject and
createUpdateStatement to make database connection and update the given
project. Add updating a project in methods loadImage, deleteImage,
dilate and makeContrast in the class ProjectModel.