feat: Add updating the projectObservableList with the list loaded from the database
Add a method executeQuery in the class DbManager that connect
to the database and call the method callQuery. Add the method
callQuery that create a SQL query and return a query result. Add the
method createCachedRowSet that execute a query and prepare a final
result. Add the method getAllProjects in the class ProjectDao that
execute the query loading all projects and return a project list. Update
the method setProjectObservableList in the class ProjectList so that
it uses the method getAllProjects to update the projectObservableList.
Add updating the list in the method deleteProject. Add the method
updateProjectList in the class ProjectModel and calling this method
in methods init and saveAcitveProject. Exchange the sourceFile data
type for String and alter methods which use it.