feat: Add protection against saving new project with an existing name
Add an exception class ProjectSaveException. Add a method isProject
in the class ProjectDao that filters result of the query to check
whether given name exists in the database or not. Add calling this method
in the method saveActiveProject in the class ProjectModel and handling
ProjectSaveException in the method saveProject in the MainProjectController.