refactor: Add image positioning in the center of viewports
Refactor the method scaleImage in the class ImageSettings, so
that it uses ImageView transforms. Add methods pressImage, dragImage,
setScrolPaneHvalue and setScrolPaneVvalue which let move image
using a mouse. Add methods pressPhoto, pressProject, dragPhoto
and dragProject in the class MainProjectController that call static
methods from the class ImageSettings. Add groups to ScrollPanes in
the class MainProjectController.