feat: Add selecting application language
Add a new file messages_de.properties with German messages.
Create a class ResourceBundleFactory in the package utils
that manages locale. Add a method setLocale which sets given locale
as default and calls a method reloadLayout. Add the method reloadLayout
that calls a static method refresh in the class App. Add the method
refresh in the class App. Add methods setPolishLanguage, setEnglishLanguage
and setGermanLanguage in the class MainMenuButtonsController that
call the method setLocale from the class ResourceBundleFactory.