feat: Add projects' overview
Create a fxml file ProjectsOverviewLayout with its controller
ProjectsOverviewController. Create classes ProjectOverview and
ProjectsOverviewModel. Add a method resize in the class App and
a method present in the class ProjectOverview. Call these methods
in the controller initialize method to create the overview layout.
Add an action event to project buttons that save selected project
in the class App and present it using MainProjectLayout. Add
checking if any project is stored during project opening. Create a
button which closes the overview.