feat: Add opening the application in the previously set size
Add storing whether the stage is maximized or not in the table "stages"
in the database. Add a method getMainStageSize in the class ProjectDao
that returns the requested value. Call it in the class Main. Add
sizing stage, depending on the stored value. Add a method
executeStageSizeUpdate in the class DbManager that update this
value and calls it in the method updateMainStageSize in the class
ProjectDao. Add updating the value in the database during switching
stage size in the class App.